Support SQL include statements
