+++ packages/language/src/language-server/completion/follow-elements.ts
@@ -220,7 +220,8 @@ const semicolonFollowKinds = new Set([
   CstNodeKind.DisplayStatement_Semicolon,
   CstNodeKind.DoStatement_Semicolon,
   CstNodeKind.EndStatement_Semicolon,
-  CstNodeKind.ExecStatement_Semicolon,
+  CstNodeKind.ExecCicsStatement_Semicolon,
+  CstNodeKind.ExecSqlStatement_Semicolon,
   CstNodeKind.ExitStatement_Semicolon,
   CstNodeKind.FetchStatement_Semicolon,
   CstNodeKind.FlushStatement_Semicolon,
