+++ packages/language/src/syntax-tree/cst.ts
@@ -270,12 +270,13 @@ export enum CstNodeKind {
   DoSpecification_UPTHRU,
   DoSpecification_DOWNTHRU,
   DoSpecification_REPEAT,
-  ExecStatement_Token,
-  ExecStatement_Semicolon,
   ExecCicsStatement_EXEC,
   ExecCicsStatement_CICS,
+  ExecCicsStatement_Semicolon,
   ExecSqlStatement_EXEC,
   ExecSqlStatement_SQL,
+  ExecSqlStatement_Semicolon,
+  EmbeddedUnknownStatement_Token,
   ExitStatement_EXIT,
   ExitStatement_Semicolon,
   FetchStatement_FETCH,
